feat: add multi-switch control capability for target
Virtual Switch Enabler 1.x could only control a single target switch. In order to control multiple switches, it was necessary to trigger a rule. VSE 2.0.0 adds the capibility to select multiple switches for control by the application, eliminating the need for an intermediate rule.

BREAKING CHANGE: Unable to import 1.x rules to 2.x, due to the fact that `targetSwitches` now contains a collection of switches, rather than a single switch.

feat/multi-switch-target